+++ src/app/np-dialog-demo/np-dialog-demo.component.ts
@@ -18,10 +18,10 @@ this.modalService.open(NpDialogComponent, null, { type: 'alert', message: 'Saved
 <span class="np-text-success">// ----For prompt dialog, pass type="prompt"</span>
 const prompt = this.modalService.open(NpDialogComponent, null, { type: 'prompt', message: 'Enter your name' });
 prompt.onClose.subscribe((data) => {
-  <span class="np-text-success">// if Ok button is clicked then returns {action : 'Ok', value: string }</span>
-  <span class="np-text-success">// if Cancel button is clicked then returns {action : 'Cancel' }</span>
-  if (data.action === 'Ok') {
-    alert('Ok button is clicked and input value is : ' + data.value);
+  <span class="np-text-success">// if Ok button is clicked then returns input value as string</span>
+  <span class="np-text-success">// if Cancel button is clicked then returns input value as null</span>
+  if (data != undefined && data != null) {
+    alert('Ok button is clicked and input value is: ' + data);
   } else {
     alert('Cancel button is clicked');
   }
@@ -30,9 +30,9 @@ prompt.onClose.subscribe((data) => {
 <span class="np-text-success">// ----For confirm dialog, pass type="confirm"</span>
 const confirm = this.modalService.open(NpDialogComponent, null, { type: 'confirm', message: 'Are you sure to delete?' });
 confirm.onClose.subscribe((data) => {
-  <span class="np-text-success">// if Ok button is clicked then returns {action : 'Ok'}</span>
-  <span class="np-text-success">// if Cancel button is clicked then returns {action : 'Cancel' }</span>
-  if (data.action === 'Ok') {
+  <span class="np-text-success">// if Ok button is clicked then returns true</span>
+  <span class="np-text-success">// if Cancel button is clicked then returns false</span>
+  if (data) {
     alert('Ok button is clicked');
   } else {
     alert('Cancel button is clicked');
@@ -54,7 +54,11 @@ confirm.onClose.subscribe((data) => {
     const prompt = this.modalService.open(NpDialogComponent,
       null, { type: 'prompt', message: 'Enter your name' });
     prompt.onClose.subscribe((data) => {
-      alert(JSON.stringify(data));
+      if (data != undefined && data != null) {
+        alert('Ok button is clicked, input value is: ' + data);
+      } else {
+        alert('Cancel button is clicked');
+      }
     });
   }
 
@@ -62,7 +66,11 @@ confirm.onClose.subscribe((data) => {
     const confirm = this.modalService.open(NpDialogComponent,
       null, { type: 'confirm', message: 'Are you sure to delete?' });
     confirm.onClose.subscribe((data) => {
-      alert(JSON.stringify(data));
+      if (data) {
+        alert('Ok button is clicked');
+      } else {
+        alert('Cancel button is clicked');
+      }
     });
   }
 }